Refactor: 데이터 패칭 SSR로 변경
